create abstract ArchiveDayRepository and extends repositories with it
